Refactor file storage operations into individual actions
Decompose monolithic File Storage Operations action into focused, single-purpose actions for improved maintainability and clarity. Each storage operation now has its own action class with a shared base class handling common functionality.

Changes:
- Split File Storage Operations into 13 individual actions (List, Get, Copy, Move, Delete, etc.)
- Create BaseFileStorageAction with shared provider lookup and error handling
- Update metadata for all new actions with proper parameters and result codes
- Update migration file with new action definitions
- Export all new actions from index file
