+++ packages/Actions/CoreActions/src/custom/files/index.ts
@@ -1,0 +1,23 @@
+/**
+ * File Storage Actions
+ *
+ * Granular actions for interacting with various cloud storage providers
+ * (Azure Blob Storage, AWS S3, Google Cloud Storage, Google Drive, Dropbox, Box.com, SharePoint).
+ *
+ * Each action performs a single, focused operation on a storage provider.
+ * All actions extend BaseFileStorageAction for shared functionality.
+ */
+
+export { BaseFileStorageAction } from './base-file-storage.action';
+export { ListObjectsAction } from './list-objects.action';
+export { GetMetadataAction } from './get-metadata.action';
+export { GetObjectAction } from './get-object.action';
+export { GetDownloadUrlAction } from './get-download-url.action';
+export { GetUploadUrlAction } from './get-upload-url.action';
+export { ObjectExistsAction } from './object-exists.action';
+export { DirectoryExistsAction } from './directory-exists.action';
+export { CopyObjectAction } from './copy-object.action';
+export { MoveObjectAction } from './move-object.action';
+export { DeleteObjectAction } from './delete-object.action';
+export { CreateDirectoryAction } from './create-directory.action';
+export { DeleteDirectoryAction } from './delete-directory.action';
