+++ packages/Actions/CoreActions/src/index.ts
@@ -47,7 +47,9 @@ export * from './custom/files/pdf-extractor.action';
 export * from './custom/files/excel-reader.action';
 export * from './custom/files/excel-writer.action';
 export * from './custom/files/file-compress.action';
-export * from './custom/files/file-storage-operations.action';
+
+// File Storage Actions - Granular operations for cloud storage
+export * from './custom/files';
 
 // Integration Actions
 export * from './custom/integration/http-request.action';
@@ -111,6 +113,18 @@ import { LoadPDFExtractorAction } from './custom/files/pdf-extractor.action';
 import { LoadExcelReaderAction } from './custom/files/excel-reader.action';
 import { LoadExcelWriterAction } from './custom/files/excel-writer.action';
 import { LoadFileCompressAction } from './custom/files/file-compress.action';
+import { LoadListObjectsAction } from './custom/files/list-objects.action';
+import { LoadGetMetadataAction } from './custom/files/get-metadata.action';
+import { LoadGetObjectAction } from './custom/files/get-object.action';
+import { LoadGetDownloadUrlAction } from './custom/files/get-download-url.action';
+import { LoadGetUploadUrlAction } from './custom/files/get-upload-url.action';
+import { LoadObjectExistsAction } from './custom/files/object-exists.action';
+import { LoadDirectoryExistsAction } from './custom/files/directory-exists.action';
+import { LoadCopyObjectAction } from './custom/files/copy-object.action';
+import { LoadMoveObjectAction } from './custom/files/move-object.action';
+import { LoadDeleteObjectAction } from './custom/files/delete-object.action';
+import { LoadCreateDirectoryAction } from './custom/files/create-directory.action';
+import { LoadDeleteDirectoryAction } from './custom/files/delete-directory.action';
 import { LoadHTTPRequestAction } from './custom/integration/http-request.action';
 import { LoadGraphQLQueryAction } from './custom/integration/graphql-query.action';
 import { LoadOAuthFlowAction } from './custom/integration/oauth-flow.action';
@@ -169,6 +183,18 @@ export function LoadAllCoreActions() {
     LoadExcelReaderAction();
     LoadExcelWriterAction();
     LoadFileCompressAction();
+    LoadListObjectsAction();
+    LoadGetMetadataAction();
+    LoadGetObjectAction();
+    LoadGetDownloadUrlAction();
+    LoadGetUploadUrlAction();
+    LoadObjectExistsAction();
+    LoadDirectoryExistsAction();
+    LoadCopyObjectAction();
+    LoadMoveObjectAction();
+    LoadDeleteObjectAction();
+    LoadCreateDirectoryAction();
+    LoadDeleteDirectoryAction();
     LoadHTTPRequestAction();
     LoadGraphQLQueryAction();
     LoadOAuthFlowAction();
